Supply additional file information when optimizing an SVG.
The SVGO library recently added a new "prefixIds" plugin, which included support for supplying additional file information to its optimize function so that ID prefixes could be created using a file's name. This will allow us to pass along this additional information from svgo-loader so that we can generate prefixes based on filename for SVGs loaded via this loader.