Supply additional file information when optimizing an SVG to allow ID prefixing based on file name (#22)
* Supply additional file information when optimizing an SVG.

The SVGO library recently added a new "prefixIds" plugin, which included support for supplying additional file information to its optimize function so that ID prefixes could be created using a file's name. This will allow us to pass along this additional information from svgo-loader so that we can generate prefixes based on filename for SVGs loaded via this loader.

* Remove `input` property when info object passed during optimization

This property isn't used anywhere in the code and is not present in SVGO unit test examples.