update webpack and svgo dependencies (#58)
* update webpack and svgo dependencies

This updates to svgo 3. Because `optimize` now throws instead of
returning an error result, error handling has been removed.

This also removes support for Webpack 4. This allows to use
`this.getOptions()` instead of `loader-utils`. Also the Webpack
configuration files have been updated for Webpack 5.

As of npm 9 the default lockfile format version is 3. This is now
explicitly specified in .npmrc for compatibility between npm 8 and 9.

* update readme for Webpack 5